Moved newly added test cases to separate file
