+++ pkg/movies/api_test.go
@@ -2,7 +2,6 @@ package movies
 
 import (
 	"net/http"
-	"sort"
 	"testing"
 
 	"github.com/jarcoal/httpmock"
@@ -53,48 +52,6 @@ func TestSearchMovies(t *testing.T) {
 			} else {
 				assert.EqualError(t, actualError, c.expectedErrorString)
 			}
-		})
-	}
-}
-
-func TestSort(t *testing.T) {
-	cases := []struct {
-		name            string
-		movies          Movies
-		expected_movies Movies
-	}{
-		{
-			name: "RegularSort",
-			movies: Movies{
-				{Title: "The Godfather: Part III", Year: "1990"},
-				{Title: "The Godfather: Part II", Year: "1974"},
-				{Title: "The Godfather", Year: "1972"},
-			},
-			expected_movies: Movies{
-				{Title: "The Godfather", Year: "1972"},
-				{Title: "The Godfather: Part II", Year: "1974"},
-				{Title: "The Godfather: Part III", Year: "1990"},
-			},
-		},
-		{
-			name: "SameYearSort",
-			movies: Movies{
-				{Title: "The Matrix Revolutions", Year: "2003"},
-				{Title: "The Matrix", Year: "1990"},
-				{Title: "The Matrix Reloaded", Year: "2003"},
-			},
-			expected_movies: Movies{
-				{Title: "The Matrix", Year: "1990"},
-				{Title: "The Matrix Reloaded", Year: "2003"},
-				{Title: "The Matrix Revolutions", Year: "2003"},
-			},
-		},
-	}
-
-	for _, c := range cases {
-		t.Run(c.name, func(t *testing.T) {
-			sort.Sort(Movies(c.movies))
-			assert.EqualValues(t, c.movies, c.expected_movies)
 		})
 	}
 }
